--- src/components/organisms/teacher/OakDropdownNavMenu/index.ts
@@ -1,1 +1,0 @@
-export * from "./OakButtonWithDropdown";
